Refactor layers, restore snake case for arguments
